+++ tests/postgres_integration.rs
@@ -1,10 +1,13 @@
 use scouter_server::sql::schema::DriftRecord;
 use sqlx::Row;
-mod common;
+
+use scouter_server::sql::postgres::PostgresClient;
+mod test_utils;
 
 #[tokio::test]
 async fn test_postgres_client() {
-    let (db_client, _) = common::setup_test_db().await.unwrap();
+    let pool = test_utils::setup_pool_and_clean_db().await.unwrap();
+    let db_client = PostgresClient::new(pool.clone()).unwrap();
 
     // test inserting record
     let record = DriftRecord {
@@ -64,5 +67,5 @@ async fn test_postgres_client() {
 
     assert_eq!(result.features.len(), 1);
 
-    common::teardown().await.unwrap();
+    test_utils::teardown().await.unwrap();
 }
